split ListenerTest methods into some new classes
